Finished adding comments to my code and changed the docs because I cant make up my mind about how the languae should work. I also corrected a few spelling mistakes
