+++ docs/commands/createStart/ADDDESC.md
@@ -5,10 +5,10 @@
 # .CreateStart.addDesc
 
 **About:**
-```.addDesc``` is used to add the descption of the story.
+```.addDesc``` is used to add the description of the story.
 
 **Usage:**
-```.<StartScene>.addDesc <Desc>```
+```<StartScene>.addDesc <Desc>```
 
 **Args:**
