+++ docs/exampleStorys/EX1.md
@@ -10,23 +10,23 @@
 **Code:**
 ```
 > Creates a start scene.
-/start = .CreateStart
+start = .CreateStart
 > Sets up the start scene with a title, description, and text.
-/start.addTitle "Example Story #1"
-/start.addDesc "This is an example story that has two scenes in it"
-/start.addText "You wake up in your bed"
+start.addTitle "Example Story #1"
+start.addDesc "This is an example story that has two scenes in it"
+start.addText "You wake up in your bed"
 
 > Creates a dead-end scene.
-/end = .CreateDeadEnd
+end = .CreateDeadEnd
 > Sets up the dead-end scene
-/end.addText "You go back to sleep"
+end.addText "You go back to sleep"
 
 > Creates a new option that is "Go back to sleep".
-/option1 = .CreateOption "Go back to sleep"
+option1 = .CreateOption "Go back to sleep"
 > Links the option to the dead-end scene.
-/option1.link end
+option1.link end
 > Adds the option to the start scene.
-/start.addOption option1
+start.addOption option1
 ```
 
 **Graph:**
