+++ docs/exampleStorys/MAIN.md
@@ -5,7 +5,7 @@
 
 # Example Stories
 
-We have provided a few exmple stories with the code and graph for you to really understand what is going on.
+We have provided a few example stories with the code and graph for you to really understand what is going on.
 
 **[Example Story #1](/exampleStorys/EX1.md):** This is a short one that demonstations normal use with a few diffrent branches.
