+++ src/Engine/other/player.java
@@ -1,8 +1,15 @@
-package Engine;
-import Engine.scene;
-import java.util.InputMismatchException;
+package Engine.other;
+import Engine.scenes.savePoint;
+import Engine.scenes.scene;
+
 import java.util.Scanner;
 
+/**
+ * The player object. This has the logic for the scenes and story. The player also stores important story data that cant be stored in the scene.
+ *
+ * @author Omar M. Radwan
+ * @version 1.0.0
+ */
 public class player {
 	private String name;
 	private scene lastScene = null;
@@ -21,24 +28,43 @@ public class player {
 		this.startScene = scene;
 		userIn = new Scanner(System.in);
 	}
-	
+
+	/**
+	 * Sets what scene we are on
+	 * @param scene - The scene we are on
+	 */
 	public void setScene(scene scene) {
 		this.lastScene = this.currScene;
 		this.currScene = scene;
 	}
 
+	/**
+	 * Gets the saved scene
+	 * @return - The saved scene
+	 */
 	public scene getSaveScene() {
 		return this.saveScene;
 	}
-	
+
+	/**
+	 * Gets the last scene
+	 * @return - The last scene
+	 */
 	public scene getLastScene() {
 		return this.lastScene;
 	}
 
+	/**
+	 * Gets the start scene
+	 * @return - The start scene
+	 */
 	public scene getStartScene() {
 		return this.startScene;
 	}
 
+	/**
+	 * The main loop of the engine. This prints the scene and options and wait for the user to choose an option and moves to the appropriate scene
+	 */
 	public void play() {
 		if (this.currScene != null) {
 			this.currScene.compileScene(this);
